CHG: Search video list functionality

+++ src/app/videos-table/videos-table.component.ts
@@ -13,11 +13,19 @@ import { ButtonType } from '../shared/types/button-type';
 import { MatDialog, MatDialogModule } from '@angular/material/dialog';
 import { VideoDeleteComponent } from '../video-delete/video-delete.component';
 import { DataService } from '../shared/data/data.service';
+import { FormsModule } from '@angular/forms';
+import { SearchPipe } from '../shared/search.pipe';
 
 @Component({
   selector: 'mi-videos-table',
   standalone: true,
-  imports: [CommonModule, ButtonComponent, MatDialogModule],
+  imports: [
+    CommonModule,
+    ButtonComponent,
+    MatDialogModule,
+    FormsModule,
+    SearchPipe
+  ],
   templateUrl: './videos-table.component.html',
   styleUrls: ['./videos-table.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush,
@@ -30,9 +38,10 @@ export class VideosTableComponent {
   dialog = inject(MatDialog);
   @Input() videos: ProcessedVideo[] = [];
   protected readonly buttonType = ButtonType;
+  searchArg = '';
 
   onEdit(video: ProcessedVideo) {
-    this.router.navigate(['edit', video.id], { relativeTo: this.activatedRoute });
+    this.router.navigate(['edit', video.id], { relativeTo: this.activatedRoute, state: video });
   }
 
   onDelete(video: ProcessedVideo) {
@@ -44,16 +53,18 @@ export class VideosTableComponent {
       data: video,
     });
 
-    dialogRef.afterClosed().subscribe(() => {
-      this.dataService.deleteVideo(video.authorId, video.id).subscribe(
-        (result) => { 
-          if(result) {
-            const index = this.videos.findIndex((videoEl) => videoEl.id === video.id);
-            this.videos.splice(index, 1);
-            this.changeDetRef.detectChanges();
+    dialogRef.afterClosed().subscribe((dialogResult) => {
+      if (dialogResult) {
+        this.dataService.deleteVideo(video.authorId, video.id).subscribe(
+          (result) => { 
+            if(result) {
+              const index = this.videos.findIndex((videoEl) => videoEl.id === video.id);
+              this.videos.splice(index, 1);
+              this.changeDetRef.detectChanges();
+            }
           }
-        }
-      );
+        );
+      }
     });
   }
 }
